+++ src/Exporter.php
@@ -11,4 +11,6 @@ interface Exporter
     public function exportProfilingData(ProfilingResult|ProfilingData $profilingData): void;
 
     public function exportProfiler(Profiler $profiler, ?string $label = null): void;
+
+    public function exportTimeline(Timeline $timeline): void;
 }