+++ src/ProfilingData.php
@@ -6,8 +6,6 @@ namespace Bakame\Aide\Profiler;
 
 use JsonSerializable;
 
-use function preg_match;
-
 /**
  * @phpstan-import-type MetricsStat from Metrics
  * @phpstan-import-type SnapshotStat from Snapshot
@@ -28,12 +26,10 @@ final class ProfilingData implements JsonSerializable
      */
     public function __construct(Snapshot $start, Snapshot $end, string $label)
     {
-        1 === preg_match('/^[a-z0-9][a-z0-9_]*$/', $label) || throw new InvalidArgument('The label must start with a lowercased letter or a digit and only contain lowercased letters, digits, or underscores.');
-
         $this->metrics = Metrics::fromSnapshots($start, $end);
         $this->start = $start;
         $this->end = $end;
-        $this->label = $label;
+        $this->label = Label::fromString($label);
     }
 
     /**